Factor out NormalizePlan to be later used in single-pass Analyzer testing
